sync:2506代码同步 84b2bc98d88c70dfa79448891f2b2e022369a4b2

+++ BSWCode/Communication/Ethernet/TcpIp/TcpIp_Internal.h
@@ -435,7 +435,7 @@ FUNC(TcpIp_StateType, TCPIP_CODE) TcpIp_GetControlState(VAR(uintx, AUTOMATIC) ct
 #endif /* TCPIP_CONTROLLER_NUM > 0 */
 
 #if (TCPIP_SOCKET_NUM > 0u)
-FUNC(void, TCPIP_CODE) TcpIp_SocketInit(VAR(uint8, AUTOMATIC) socketNum);
+FUNC(void, TCPIP_CODE) TcpIp_SocketInit(VAR(uint16, AUTOMATIC) socketNum);
 #endif /* TCPIP_SOCKET_NUM > 0u */
 
 #if (TCPIP_LOCAL_ADR_NUM > 0u)
